pyproject.toml
[tool.poetry]
name = "comic-gen-flux"
version = "0.1.0"
description = ""
authors = ["Tímea Tóth <tiemka14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.3.0"
loguru = "^0.7.3"
ruff = "^0.12.3"
black = "^25.1.0"
rich = "^14.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
